MO: Clean up videoio __init__.py
